refactor update delete service

+++ src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, InternalServerErrorException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -22,18 +22,40 @@ export class UserService {
   }
 
   public async findOne(id: number) {
-    return await this.userRepository.findOne({
+    const user = await this.userRepository.findOne({
       where: {
         id
       }
     })
+
+    if (!user)
+      throw new HttpException(
+        'User does not exist',
+        HttpStatus.BAD_REQUEST,
+      );
+
+    return user
   }
 
-  public async update(input: UpdateUserDto) {
-    return await this.userRepository.save(input)
+  public async update(id: number, input: UpdateUserDto) {
+    const user = await this.findOne(id)
+
+    if (!user) {
+      throw new InternalServerErrorException()
+    }
+
+    const mergeUser = this.userRepository.merge(user, input)
+
+    return this.userRepository.save(mergeUser)
   }
 
   public async remove(id: number) {
+    const user = await this.findOne(id)
+
+    if (!user) {
+      throw new InternalServerErrorException()
+    }
+
     return await this.userRepository.delete(id)
   }
 }
